requirements.txt
absl-py==2.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
attrs==26.1.0
blinker==1.9.0
bokeh==3.9.0
cachetools==7.0.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fastapi==0.136.0
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-auth-oauthlib==1.3.1
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==2.15.0
kiwisolver==1.5.0
libclang==18.1.1
livelossplot==0.5.6
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
ml-dtypes==0.2.0
narwhals==2.20.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==26.1
pandas==3.0.2
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
protobuf==4.25.9
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.8
wrapt==1.14.2
xyzservices==2026.3.0
